+++ sketch-extensions-apkicon/src/androidTest/kotlin/com/github/panpf/sketch/extensions/apkicon/test/decode/internal/ApkIconDecoderComponentTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.extensions.apkicon.test.decode.internal
+
+class ApkIconDecoderComponentTest {
+    // TODO test
+}