+++ sketch-extensions-appicon/src/androidTest/kotlin/com/github/panpf/sketch/extensions/appicon/test/drawable/AppIconDrawableFetcherTest.kt
@@ -1,4 +1,4 @@
-package com.github.panpf.sketch.extensions.core.android.test.drawable
+package com.github.panpf.sketch.extensions.appicon.test.drawable
 
 import com.github.panpf.sketch.drawable.AppIconDrawableFetcher
 import com.github.panpf.sketch.test.utils.getTestContext