+++ sketch-extensions-appicon/src/androidTest/kotlin/com/github/panpf/sketch/extensions/appicon/test/fetch/internal/AppIconUriFetcherComponentTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.extensions.appicon.test.fetch.internal
+
+class AppIconUriFetcherComponentTest {
+    // TODO test
+}