+++ sketch-extensions-core/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/extensions/core/android/test/util/ExtensionsCoreAndroidUtilsTest.kt
@@ -2,10 +2,8 @@ package com.github.panpf.sketch.extensions.core.android.test.util
 
 import android.content.res.Resources
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.panpf.sketch.test.utils.getTestContext
 import com.github.panpf.sketch.util.dp2Px
 import com.github.panpf.sketch.util.format
-import com.github.panpf.sketch.util.versionCodeCompat
 import org.junit.runner.RunWith
 import kotlin.math.roundToInt
 import kotlin.test.Test
@@ -40,16 +38,5 @@ class ExtensionsCoreAndroidUtilsTest {
             expected = (3.33f * Resources.getSystem().displayMetrics.density + 0.5f).roundToInt(),
             3.33f.dp2Px()
         )
-    }
-
-    @Test
-    fun testVersionCodeCompat() {
-        val context = getTestContext()
-        val packageInfo = context.packageManager.getPackageInfo(context.packageName, 0)
-        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.P) {
-            assertEquals(packageInfo.longVersionCode.toInt(), packageInfo.versionCodeCompat)
-        } else {
-            assertEquals(packageInfo.versionCode, packageInfo.versionCodeCompat)
-        }
     }
 }