+++ sketch-extensions-core/src/androidMain/kotlin/com/github/panpf/sketch/util/extensions_core_android_utils.kt
@@ -16,7 +16,6 @@
 
 package com.github.panpf.sketch.util
 
-import android.content.pm.PackageInfo
 import kotlin.math.pow
 import kotlin.math.round
 import kotlin.math.roundToInt
@@ -24,7 +23,7 @@ import kotlin.math.roundToInt
 /**
  * Format the number to the specified number of decimal places
  *
- * @see com.github.panpf.sketch.extensions.core.android.test.internal.ExtensionsCoreAndroidUtilsTest.testFormat
+ * @see com.github.panpf.sketch.extensions.core.android.test.util.ExtensionsCoreAndroidUtilsTest.testFormat
  */
 internal fun Float.format(newScale: Int): Float {
     return if (this.isNaN()) {
@@ -38,21 +37,8 @@ internal fun Float.format(newScale: Int): Float {
 /**
  * Convert dp to px
  *
- * @see com.github.panpf.sketch.extensions.core.android.test.internal.ExtensionsCoreAndroidUtilsTest.testDp2Px
+ * @see com.github.panpf.sketch.extensions.core.android.test.util.ExtensionsCoreAndroidUtilsTest.testDp2Px
  */
 internal fun Float.dp2Px(): Int {
     return (this * android.content.res.Resources.getSystem().displayMetrics.density + 0.5f).roundToInt()
 }
-
-/**
- * Get the version code that is compatible with all Android versions
- *
- * @see com.github.panpf.sketch.extensions.core.android.test.internal.ExtensionsCoreAndroidUtilsTest.testVersionCodeCompat
- */
-@Suppress("DEPRECATION")
-internal val PackageInfo.versionCodeCompat: Int
-    get() = if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.P) {
-        longVersionCode.toInt()
-    } else {
-        versionCode
-    }